backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0

# Cache
redis==5.0.1

# AWS
boto3==1.34.25

# Auth
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Config
pyyaml==6.0.1

# Logging
python-json-logger==2.0.7

# HTTP client
httpx==0.26.0

# Dev/Test
pytest==7.4.4
pytest-asyncio==0.23.3
ruff==0.1.13
